+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Option to download all annotations.
+
 ### Fixed
 
 - Handle `purple` and `dark_blue` mark colors.